requirements.txt
pandas==2.0.3
openai==1.55.2
numpy==1.23.2
python-dotenv==1.0.1
mlxtend==0.23.4
firebase-admin==6.5.0
google-cloud-storage==3.0.0
langchain-community
streamlit
faiss-cpu
boto3